refactor(move-package): bet관련 테스트 패키지 이동
bet과 관련된 테스트의 패키지를 bet 내부로 이동했습니다.
