+++ src/test/java/domain/bet/BetHistoryTest.java
@@ -1,9 +1,8 @@
-package domain;
+package domain.bet;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
-import domain.bet.BetHistory;
 import domain.participant.Name;
 import java.util.List;
 import java.util.Map;
