+++ src/test/java/domain/bet/BetProfitTest.java
@@ -1,8 +1,7 @@
-package domain;
+package domain.bet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.bet.BetProfit;
 import domain.enums.GameResult;
 import domain.participant.Name;
 import java.util.List;
